originalPrice and OrderUrl in wide, blog author fix
